requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
httpx
pydantic
python-multipart
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0